+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## v4.1.2 (2022-07-20)
+* minicap kill previous instance on startup
+* minicap only upload its binary if thet are missing
 * fix DeviceClient.tryForwardTCP failing with multiple devices
 
 ## v4.1.1 (2022-07-20)